# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentFeeStatus =
  (typeof StudentFeeStatus)[keyof typeof StudentFeeStatus];

export const StudentFeeStatus = {
  paid: "paid",
  partial: "partial",
  pending: "pending",
  unknown: "unknown",
} as const;

export type StudentLanguagePreference =
  (typeof StudentLanguagePreference)[keyof typeof StudentLanguagePreference];

export const StudentLanguagePreference = {
  english: "english",
  kannada: "kannada",
  hindi: "hindi",
  telugu: "telugu",
  tamil: "tamil",
} as const;

export type StudentRiskLevel =
  (typeof StudentRiskLevel)[keyof typeof StudentRiskLevel];

export const StudentRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Student {
  id: number;
  name: string;
  phone: string;
  college: string;
  branch: string;
  year: number;
  /** @nullable */
  district: string | null;
  /** @nullable */
  incomeRange: string | null;
  /** @nullable */
  casteCategory: string | null;
  feeStatus: StudentFeeStatus;
  languagePreference: StudentLanguagePreference;
  onboardingComplete: boolean;
  onboardingStep?: number;
  riskScore: number;
  riskLevel: StudentRiskLevel;
  daysSilent: number;
  /** @nullable */
  lastMessageAt?: string | null;
  flagged?: boolean;
  createdAt: string;
  updatedAt: string;
}

export type StudentUpdateFeeStatus =
  (typeof StudentUpdateFeeStatus)[keyof typeof StudentUpdateFeeStatus];

export const StudentUpdateFeeStatus = {
  paid: "paid",
  partial: "partial",
  pending: "pending",
  unknown: "unknown",
} as const;

export type StudentUpdateRiskLevel =
  (typeof StudentUpdateRiskLevel)[keyof typeof StudentUpdateRiskLevel];

export const StudentUpdateRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface StudentUpdate {
  feeStatus?: StudentUpdateFeeStatus;
  flagged?: boolean;
  riskScore?: number;
  riskLevel?: StudentUpdateRiskLevel;
}

export type ConversationMessageDirection =
  (typeof ConversationMessageDirection)[keyof typeof ConversationMessageDirection];

export const ConversationMessageDirection = {
  inbound: "inbound",
  outbound: "outbound",
} as const;

export interface ConversationMessage {
  id: number;
  studentId: number;
  direction: ConversationMessageDirection;
  message: string;
  /** @nullable */
  agentType: string | null;
  createdAt: string;
}

export type RiskFlagSeverity =
  (typeof RiskFlagSeverity)[keyof typeof RiskFlagSeverity];

export const RiskFlagSeverity = {
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface RiskFlag {
  id: number;
  studentId: number;
  studentName: string;
  /** @nullable */
  studentCollege?: string | null;
  severity: RiskFlagSeverity;
  reason: string;
  riskScore: number;
  resolved: boolean;
  /** @nullable */
  resolvedAt?: string | null;
  createdAt: string;
}

export interface Scholarship {
  id: number;
  name: string;
  amount: number;
  deadline: string;
  eligibilitySummary: string;
  category: string;
  isActive?: boolean;
}

export type DashboardSummaryRiskBreakdown = {
  low: number;
  medium: number;
  high: number;
  critical: number;
};

export interface DashboardSummary {
  totalStudents: number;
  atRiskToday: number;
  criticalCount: number;
  nudgesSentToday: number;
  scholarshipsAssisted: number;
  onboardingComplete: number;
  averageRiskScore: number;
  riskBreakdown: DashboardSummaryRiskBreakdown;
}

export interface CheckInInput {
  message?: string;
}

export interface CheckInResult {
  success: boolean;
  message: string;
  /** @nullable */
  sid?: string | null;
}

export interface TwilioWebhookInput {
  From?: string;
  Body?: string;
  MessageSid?: string;
  To?: string;
}

export type ListStudentsParams = {
  riskLevel?: ListStudentsRiskLevel;
  college?: string;
  search?: string;
  flagged?: string;
};

export type ListStudentsRiskLevel =
  (typeof ListStudentsRiskLevel)[keyof typeof ListStudentsRiskLevel];

export const ListStudentsRiskLevel = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export type ListRiskFlagsParams = {
  resolved?: string;
  severity?: ListRiskFlagsSeverity;
};

export type ListRiskFlagsSeverity =
  (typeof ListRiskFlagsSeverity)[keyof typeof ListRiskFlagsSeverity];

export const ListRiskFlagsSeverity = {
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;