# lib/api-zod/src/generated/types/checkInResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */

export interface CheckInResult {
  success: boolean;
  message: string;
  /** @nullable */
  sid?: string | null;
}