# lib/api-zod/src/generated/types/conversationMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */
import type { ConversationMessageDirection } from "./conversationMessageDirection";

export interface ConversationMessage {
  id: number;
  studentId: number;
  direction: ConversationMessageDirection;
  message: string;
  /** @nullable */
  agentType: string | null;
  createdAt: string;
}