# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryRiskBreakdown } from "./dashboardSummaryRiskBreakdown";

export interface DashboardSummary {
  totalStudents: number;
  atRiskToday: number;
  criticalCount: number;
  nudgesSentToday: number;
  scholarshipsAssisted: number;
  onboardingComplete: number;
  averageRiskScore: number;
  riskBreakdown: DashboardSummaryRiskBreakdown;
}