# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */
import type { StudentFeeStatus } from "./studentFeeStatus";
import type { StudentLanguagePreference } from "./studentLanguagePreference";
import type { StudentRiskLevel } from "./studentRiskLevel";

export interface Student {
  id: number;
  name: string;
  phone: string;
  college: string;
  branch: string;
  year: number;
  /** @nullable */
  district: string | null;
  /** @nullable */
  incomeRange: string | null;
  /** @nullable */
  casteCategory: string | null;
  feeStatus: StudentFeeStatus;
  languagePreference: StudentLanguagePreference;
  onboardingComplete: boolean;
  onboardingStep?: number;
  riskScore: number;
  riskLevel: StudentRiskLevel;
  daysSilent: number;
  /** @nullable */
  lastMessageAt?: string | null;
  flagged?: boolean;
  createdAt: string;
  updatedAt: string;
}