# lib/api-zod/src/generated/types/twilioWebhookInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Disha — AI Companion for First-Generation College Students
 * OpenAPI spec version: 0.1.0
 */

export interface TwilioWebhookInput {
  From?: string;
  Body?: string;
  MessageSid?: string;
  To?: string;
}